Customsets

See merge request awa111/CMPT276-1191E1-Sigma!72

+++ app/src/main/java/com/sigma/sudokuworld/persistence/db/entities/Language.kt
@@ -12,7 +12,7 @@ import android.arch.persistence.room.PrimaryKey
 
 @Entity(tableName = "languages")
 data class Language (
-    @PrimaryKey(autoGenerate = true) var languageID: Int,
+    @PrimaryKey(autoGenerate = true) var languageID: Long,
     var name: String,
     var code: String
 )
