+++ app/src/main/java/com/sigma/sudokuworld/persistence/db/entities/Set.kt
@@ -5,7 +5,7 @@ import android.arch.persistence.room.PrimaryKey
 
 @Entity(tableName = "sets")
 data class Set (
-    @PrimaryKey(autoGenerate = true) var setID: Int,
+    @PrimaryKey(autoGenerate = true) var setID: Long,
     var name: String,
     var description: String
 )
